requirements.txt
pandas
requests
python-dotenv
pytest
pytest-mock
